Renamed bucket and module name
