+++ exasol_script_languages_developer_sandbox/cli/commands/setup_vm_bucket.py
@@ -3,7 +3,7 @@ from exasol_script_languages_developer_sandbox.cli.common import add_options
 from exasol_script_languages_developer_sandbox.cli.options.aws_options import aws_options
 from exasol_script_languages_developer_sandbox.cli.options.logging import logging_options, set_log_level
 from exasol_script_languages_developer_sandbox.lib.aws_access import AwsAccess
-from exasol_script_languages_developer_sandbox.lib.vm_bucket import run_setup_vm_bucket
+from exasol_script_languages_developer_sandbox.lib.vm_slc_bucket import run_setup_vm_bucket
 
 
 @cli.command()
